Rename predicate calculus axioms (without moving them) per 20-Dec-2018
On 20-Dec-2018, the predicate calculus axiom schemes were renumbered to
eliminate gaps caused by obsolete or redundant axiom schemes, as well
as to order them according to their current organization. Older papers,
slide presentations, and websites may reference the old numbering.
The renumbering is shown in:
http://us.metamath.org/mpeuni/mmset.html#oldaxioms

This commit renames the axioms in the book, *without* moving them so
that it is easier to see exactly what names changed.

This is a first among a number of steps to update the metamath book so
that the book and actual set.mm file correspond.  We rename the axioms
*first*, in place, to ensure that all later changes will consistently
use the new names.  This will also make it easier to notice when things
are not in the new (intended) order.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>