Fix German reference per Ken Kubota, fixes #10
Fix a German reference.  Ken Kubota is a German native speaker,
and noticed that:

* Jahresberichte der Deutschen Mathematicker Vereinigung, bk. 2

Should be:

* Jahresbericht der Deutschen Mathematiker-Vereinigung, vol. 2, p. 19.

The explanation is as follows:
The "c" is clearly a spelling mistake and has to be removed.
(The German "k" replaces the English "c".)
The hyphen ("-") is part of the journal's title.
The "e" signals plural (like the "s" in "annual reports") and should be
removed, as not the collection, but only a single report is referenced.
The exact location is: vol. 2, p. 19. (There the quote is provided in a
text by H. Weber. See this link (encoded);
<https://en.wikipedia.org/w/index.php?title=3DLeopold_Kronecker&oldid=3D81=
9095772#cite_note-1> at Wikipedia and the original text here =
<http://www.digizeitschriften.de/dms/img/?PID=3DPPN37721857X_0002%7Clog6&p=
hysid=3Dphys10#navi>, p. 23:19.)

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>