Update the axiom short list in 3.3
Update section 3.3 "The axioms in the Metamath Language" to list
the *actual* axioms now used in set.mm.  Basically it's Tarski's S2 and
a few auxiliaries.

This commit does *not* fix all the text to match.
This just tries to get the correct axioms in the correct order; once
they're in, we can modify the various surrounding text to match.
Note that ax-5 is modified so that the condition is inline.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>