Add the ability to generate metamath-narrow.pdf for smartphones etc.
Today *many* people have smartphones, but the traditional wide
formatting doesn't work so well on them.  PDF supports reflowing,
but LaTeX tools don't support generating PDFs with varying width
so well.

Simple solution: make it easy to create metamath-narrow.pdf, which
is the same document formatted for a narrower screen.
This is a partial extraction from kindle-changes.patch, though we
currently don't do everything it did.  In the future we may extract
more, or do more things; this is simply a minimal version that makes
things work.

There may be other ways to solve this long-term, but this "gets it done".

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>