Check metamath.log for unexpected errors during generation
Modify generate-pdf so that it
checks metamath.log for unexpected errors and warnings during generation.
There are some errors and warnings we *do* expect;
those are listed (and ignored) in allowed-errors.
If there is an unexpected error or warning, then generate-pdf will
return false (failure).

We only check on the *last* generation of the PDF.
While this slightly delays error reporting in some cases,
it's done this way to eliminate the risk of false positives caused
by erroring out too soon.

Note that since our continuous integration (CI) system uses generate-pdf,
this has the effect of detecting more errors in CI.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>