Switch to Creative Commons CC0 license
The book had been previously released to the public domain using
Creative Commons Public Domain Dedication, see:
http://creativecommons.org/licenses/publicdomain/

However, Creative Commons has retired this legal tool and
does not recommend that it be applied to works.
It was US-specific and co-mingled two very different use cases.
Creative Commons have replaced it with two different tools, and the one
that matches our intent is the CC0 Public Domain Dedication.

So this commit switches the license to the
Creative Commons CC0 1.0 Universal (CC0 1.0) Public Domain Dedication per
<https://creativecommons.org/publicdomain/zero/1.0/>.

There is no *legal* requirement to do so, but we still ask that
people to give us credit (e.g., use citations).

The intent is unchanged, it's intended to be released to the
public domain.  We're just using the best available and widely accepted
legal tool to accomplish that.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>