Add a section showing 2+2=4 (2p2e4)
Add section showing 2+2=4 (2p2e4) as a LaTeX table.

I find that the proof of 2+2=4 is very helpful in understanding
Matamath, and it's one of the first things discussed on the
Metamath Proof Explorer (set.mm) front page.

Oddly, however, the proof that 2+2=4 is never shown in the book.
This commit fixes that omission.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>