Add a new section on deduction
This adds a new section on deduction, including a discussion of the
standard deduction theorem, the weak deduction theorem,
deduction form, and natural deduction.

It also adds "syl" to the list of theorems, because we refer to it.
Syl is the most commonly-used theorem, so we should mention it anyway.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>