Create new "mmraw" environment for raw Metamath text
This creates a new environment to better format text.

Many definitions and theorem fragments show raw Metamath text,
e.g., df-bi shows a "pretty" format and then the raw text.
Up to this point these have been "verbatim", but verbatim
is the wrong thing, especially in the narrow format
(where text ends up lost beyond the right edge).
It also forces line breaks in the wrong places.

Instead, create a new "mmraw" environment for this kind
of text, and set its "parindent" (first line indent) and
"hangindent" (indent for rest of paragraph) so it will line
up with the text above.  Because it's really just normal
word flow, but flushleft (aka raggedright), it will naturally
use the line length available and go to the next line
when it needs to do so.  It is expressly set to be tt font.

By setting up the environment once, we make it easy to
have consistent formatting and it works much better.

This commit only makes change in one place, df-bi,
which didn't work well in the narrow version.
If this is acceptable, then we should switch the other
similar situations.  Many other places
(e.g., df-an and df-3an) are especially badly formatted.
But there's no point in making those changes unless
it's okay to create this new environment.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>