Change "Before you start" text per #94
Thomas Brendan Leahy disagreed with the paragraph
"Before you start, you should write down on a piece of paper the complete
formal proof..." saying that,
"It led me so, so astray back in 2017".
Change it so that it's clearly Norm's experience, and that
other people find different approaches more helpful to them.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>