Move credit into hidden comment
Benoit Jubin said
"I would remove
'inspired by a request (and a lot of initial work) by Benoit Jubin'.
Actually, I would not mind if that paragraph was removed altogether:
DAW and MC already appear as overall contributors of the book, and my
contribution is in contrast negligible."

It wasn't negligible, but obviously credit is not especially desired.

Simple solution: Turn this into a comment, so it's not
included in the printed text, but we at least have some idea
of where it came from.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>