Make single quotes straight in wide version verbatim
In the wide (normal) width, display single quotes as
straight inside the verbatim environments.
Ideally we'd do that in the narrow case too, but we
can't be sure we have the right fonts.  So document
what WOULD do it, if we had the fonts, and let it be.

This at gives us straight single quotes in the
"normal" (wide) case, and I think that's an improvement.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>