Use en dash to connect multi-person names, fixes #98
It is customary to use an en dash (--) to "connect" names of different
people (and to denote ranges), and use a hyphen (-) for a
single compound name.

Fix the text so we use the en dash where needed.
Examples of connected multiple people are
Zermelo--Fraenkel, Schr\"{o}der--Bernstein, Tarski--Grothendieck,
Hewlett--Packard, and Backus--Naur.  Examples of a single person with
a compound name include Levi-Civita, Mittag-Leffler, and Burali-Forti.

Also, document this near the front so it'll be more obvious next time.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>