Fix narrow generation font problem
The previous changes accidentally required
fonts that may not be available when generating
the narrow form.

This commit modifies the generation process so that
we *always* use a special style file, either "wide.sty"
(normal case) or "narrow.sty" (narrow case).
That way we can *force* the system so that we
*only* run \usepackage{upquote} in wide mode
(this causes straight quotes in verbatim environments).
The problem is that trying to do this in narrow mode
requires fonts we may not have.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>