Improve discussion of df-sb
Add an explicit example to make clear what is being substituted for what.
Also, reorder the explanation to make it clearer.
Basically, describe what substitution does *first*, then why we
use that notation as opposed to some others, and only *then*
discuss a technical trick that the definition uses.
A reader will have a harder time understanding the technical trick if he
doesn't first understand what the definition is trying to accomplish.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>