Tweak metamath.tex to avoid problems in 11pt
Most of the book is relatively insensitive to font size.
However, if 11pt is used there were a few odd problems.

This commit modifies the text a little bit so it looks
better in 11pt.  It's mostly just additional directives
to deal with line breaks or page breaks where needed.
I did shorten a chapter starting quote,
because otherwise it goes onto a second page and the
text I elided really wasn't necessary.
I also tweaked the EBNF text so it wouldn't run off
the side of the page.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>