Circa pg.130 changes, esp. "inside of" -> "inside"
These are some small changes.

The "big" change is changing many "inside of" to "inside".
Some people assert that "inside of XXX" is the *noun*,
while "inside XXX" are the others (adjective, etc.).
In American English "inside of" is often used as a synonym,
but we don't need extra words.  So remove the excess "of"s;
they may annoy some people, and it removes unnecessary
words (yay).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>